OpenShift prefers underscores, apparently
And hates hyphens. I'm okay with changing it, to be more consistent with other env vars (although having a different convention for your own variables can be nice), but I'm less okay with *having* to change it, because it says so, and especially if it just says "Please enter a valid key. "